tls: hide an implementation detail in connection.c
From the standpoint of an external API the interesting question is
"is client certificate auth enabled or not?".  We don't need to know
about exactly which enum value to pass to gnutls to do that.

Change this to a bool.

Some (*ahem*) might argue that using the enum improves readability, but
as long as we're using clear variable names, the readability doesn't
suffer.  Also: it's not like C actually gives us any robust
type-checking between bools, ints, and enum values, so we really don't
lose anything here.