webpack-jumpstart.yml: add separate "head" build
webpack-jumpstart has been telling a lie for the past while: we've been
building the "merge" commit but labelling it in the cache with the
"head" sha.  This occasionally results in some errors about
inconsistencies, particularly in the case where package-lock.json has
changed, and is preventing us from introducing more robust consistency
checking in #16005.

The reason we do this is in large part a crutch to get around the fact
that packit only knows how to test the head commit.  Our current "lies"
allow us to get some kind of "partial merge" into packit for testing.

Meanwhile, however, packit is working on switching over to building the
merge commit instead of the head commit, and it seems like this will not
be configurable.  When that happens, our current setup will break,
because packit will go looking in our cache based on the merge commit,
but it won't find it.

Let's stop labelling merge commit build results with the head commit
sha.  Instead, we'll label them with the correct sha: the sha of the
merge commit.  That way, when packit is changed, it will continue to
work for us.

We also add an additional build step to explicitly build the "head"
commit and correctly label it as the head commit.  This temporary
measure is gated by an `if:` and will only occur if the branch isn't
fully rebased at the time of the synchronize event (and won't occur at
all for the push event).  This will allow packit to continue to work
with its current practice of checking out the head commit, and will
allow for a smooth transition when they flip the switch over to using
the merge commit.  These extra steps can be removed at that point.