bridge.py: start a cockpit-bridge equivalent, in Python
The goal is to write as much of a replacement as possible for
cockpit-bridge using nothing but the Python 3 standard library.

Excluded from this goal is the 'cockpitpackages' functionality (although
some of it is here in a hacky state, to help get things going).

The motivation is to be able to send a single Python file to any given
server and execute it in order to admin that server using Cockpit.  That
would be possible as part of Cockpit Client, for example.  The actual
html/js/css from the packages would be served directly from cockpit-ws
(since it's a little bit silly to send them all to the server, just to
have the server send them back again).