bridge.py: async!
...this was harder than it should have been.

Python's asyncio code is really strongly built around the idea of
working with sockets, so add an adaptor class that converts from
stdin/stdout (regardless of if they are a socket, pipe, terminal, file,
or anything else) to a socketpair, using threads.

Exception handling was another difficulty: any unhandled exceptions that
bubble back up to the mainloop are dealt with internally via an
exception handler, or for lack of that, being reported at the debug
level via the logging module.  Adjust our use of the logging module to
make sure the messages end up on the terminal when running locally for
testing.  Fortunately (and reasonably) exceptions result in the
connection being closed.  A feature in the stdio forwarding code to
detect the connection being closed and exiting will help with that.

Finally, it seems that although asyncio defines a BufferedProtocol type,
the default implementation of the event loop provided by the Python
standard library doesn't support it.  I wrote an efficient
implementation of the protocol based on this, but since we can't use it
without uvloop, set it aside for now (buffered.pyfragment) and use a
naive implementation.  We can tweak this in the future.